Export block settings panel layout from documents
